+++ TheForest/TheForest/Vector2.cpp
@@ -16,7 +16,7 @@ Vector2::Vector2()
 	y = 0;
 }
 
-Vector2::Vector2(int X, int Y)
+Vector2::Vector2(float X, float Y)
 {
 	x = X;
 	y = Y;
